Zabiray

+++ srcs/ft_memset.c
@@ -3,19 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_memset.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ojessi <marvin@42.fr>                      +#+  +:+       +#+        */
+/*   By: ojessi <ojessi@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2019/05/07 14:04:51 by ojessi            #+#    #+#             */
-/*   Updated: 2019/05/07 14:05:36 by ojessi           ###   ########.fr       */
+/*   Created: 2019/04/04 18:14:26 by ojessi            #+#    #+#             */
+/*   Updated: 2019/05/20 20:59:03 by ojessi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "ft_printf.h"
 
-void	*ft_memset(void *memptr, int value, size_t num)
+void	*ft_memset(void *memptr, int value, int num)
 {
 	unsigned char *str;
-
 
 	str = memptr;
 	while (num-- > 0)
